# Conflicts:
#	kalium

+++ app/src/main/kotlin/com/wire/android/ui/home/conversations/usecase/GetConversationMessagesFromSearchUseCase.kt
@@ -64,7 +64,7 @@ class GetConversationMessagesFromSearchUseCase @Inject constructor(
                 searchQuery = searchTerm,
                 conversationId = conversationId,
                 pagingConfig = pagingConfig,
-                startingOffset = max(0, lastReadIndex - PREFETCH_DISTANCE)
+                startingOffset = max(0, lastReadIndex - PREFETCH_DISTANCE).toLong()
             ).map { pagingData ->
                 pagingData.flatMap { messageItem ->
                     observeMemberDetailsByIds(messageMapper.memberIdList(listOf(messageItem)))
